requirements.txt
numpy~=1.21.4
paddlepaddle~=2.2
pgl==2.1.5
setuptools~=58.0.4
pandas~=1.4.1